Refactor directory structure
Created the following directories and moved all files to their correct
destination.

- build (Note that we are not building to ./dist)
	Scripts and tools for building, packaging and deployment.
- pkg
	Shared code that multiple services depend upon.
- srv
	Contains a subdirectory for each service. Services must be
	independent of each other and communicated using a predefined
	 API.
- test
	For mocks and other common test setup or integration tests.
- tmp
	A place for our temporary POC code.
