+++ .gitignore
@@ -33,10 +33,6 @@ bower_components
 
 # node-waf configuration
 .lock-wscript
-
-# Compiled binary addons (https://nodejs.org/api/addons.html)
-build/Release
-
 # Dependency directories
 node_modules/
 jspm_packages/
@@ -115,6 +111,6 @@ dist
 .yarn/build-state.yml
 .pnp.*
 
-build/
+dist/
 coverage/
 src/srv/bcoin/types/
