+++ jest.config.js
@@ -150,7 +150,7 @@ module.exports = {
   // ],
 
   // An array of regexp pattern strings that are matched against all test paths, matched tests are skipped
-  testPathIgnorePatterns: ["/build/", "/node_modules/"],
+  testPathIgnorePatterns: ["/dist/", "/node_modules/"],
 
   // The regexp pattern or array of patterns that Jest uses to detect test files
   // testRegex: [],
