+++ srv/bcoin/bcoin.d.ts
