+++ srv/bcoin/gettypes.sh
