+++ tmp/animal.ts
